apps/api/pyproject.toml
[tool.poetry]
name = "acog-api"
version = "0.1.0"
description = "ACOG Backend API - Automated Content Orchestration & Generation Engine"
authors = ["Justin Nichols"]
readme = "README.md"
packages = [{include = "acog", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"

# Web Framework
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Database
sqlalchemy = "^2.0.25"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"

# Task Queue
celery = {extras = ["redis"], version = "^5.3.0"}
redis = "^5.0.0"

# Object Storage
boto3 = "^1.34.0"

# AI/LLM
openai = "^1.10.0"

# HTTP Client
httpx = "^0.26.0"

# Auth & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"

# Utilities
python-dotenv = "^1.0.0"
structlog = "^24.1.0"
email-validator = "^2.3.0"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.26.0"
factory-boy = "^3.3.0"

# Linting & Formatting
ruff = "^0.2.0"
mypy = "^1.8.0"

# Type stubs
types-redis = "^4.6.0"
types-passlib = "^1.7.7"
boto3-stubs = {extras = ["s3"], version = "^1.34.0"}

# Development
watchdog = "^4.0.0"

[tool.poetry.extras]
worker = ["watchdog"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["acog"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG"]

# =============================================================================
# MyPy Configuration
# =============================================================================
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["celery.*", "boto3.*", "botocore.*"]
ignore_missing_imports = true

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
addopts = "-v --tb=short"

# =============================================================================
# Coverage Configuration
# =============================================================================
[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
